rename metric_v2 to unit_metric
